Allow configurable seed for lattice
Shuffle sharding specifically intends to use the seed used for
"randomness" to allow consistent results to be produced by the same
application.

This patch introduces a new NewLatticeWithSeed function that allows
setting the seed for the initial lattice, which will pass the seed down
to new lattices created by the SimpleShuffleShard function.

For backward compatibility the NewLattice function passes the package
level seed initialized at startup, thus users using NewLattice are
unaffected, but users can in the future choose to set the seed
themselves.